refactor(tui): move history tool run grouping
Split tool-run detection and summary helpers out of the large history renderer into history/tool_run.rs while keeping the crate::tui::history re-exports stable for existing call sites.

Refs #3308.
